refine SqlSplitterIterator
