--- server/odc-core/src/test/java/com/oceanbase/odc/core/sql/split/OBOracleSqlSplitterIteratorTest.java
@@ -1,34 +1,0 @@
-/*
- * Copyright (c) 2023 OceanBase.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.oceanbase.odc.core.sql.split;
-
-import org.antlr.v4.runtime.Lexer;
-
-import com.oceanbase.tools.sqlparser.oboracle.PLLexer;
-
-/**
- * @author gaoda.xy
- * @date 2023/11/27 20:46
- */
-public class OBOracleSqlSplitterIteratorTest extends AbstractSqlSplitterIteratorTest {
-
-    @Override
-    protected Class<? extends Lexer> lexerType() {
-        return PLLexer.class;
-    }
-
-}
