+++ server/odc-core/src/test/resources/sql/split/sql-splitter-23-multi-lines-comment-inside.yml
@@ -1,0 +1,48 @@
+origin: |-
+  select '''', '  2 '' ''/*' /* dsfgsdvcb ; ;
+  sdsd ; ; ;
+  asda*/
+  from dual;
+  select/*  
+  
+  1
+  2
+  3
+  4
+  5
+  6
+  7
+  8
+  9
+  
+  0
+  
+  1234567890*/ 
+  1 from dual;
+
+
+expected:
+  - |-
+    select '''', '  2 '' ''/*' /* dsfgsdvcb ; ;
+    sdsd ; ; ;
+    asda*/
+    from dual;
+  - |-
+    select/*  
+    
+    1
+    2
+    3
+    4
+    5
+    6
+    7
+    8
+    9
+    
+    0
+    
+    1234567890*/
+    1 from dual;
+
+expected_end_delimiter: ;
